requirements.txt
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
python-multipart~=0.0.19
pyautogen~=0.2
openpyxl==3.1.5
pandas==2.2.3